Restructured package. We now have folders for adapters, bottomlistables, fragments, and builders.
