+++ app/src/main/java/com/example/newmapsapp/ExampleClasses.java
@@ -1,5 +1,7 @@
 package com.example.newmapsapp;
 
+import com.example.newmapsapp.bottomlistable.*;
+
 import java.util.Arrays;
 import java.util.Comparator;
