+++ app/src/main/java/com/example/newmapsapp/builder/TransferPointBuilder.java
@@ -1,4 +1,9 @@
-package com.example.newmapsapp;
+package com.example.newmapsapp.builder;
+
+import com.example.newmapsapp.bottomlistable.Location;
+import com.example.newmapsapp.bottomlistable.Path;
+import com.example.newmapsapp.bottomlistable.Route;
+import com.example.newmapsapp.bottomlistable.TransferPoint;
 
 import java.util.ArrayList;
 import java.util.Arrays;
