+++ app/src/main/res/layout/destination_layout.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".DestinationLayoutFragment">
+    tools:context=".fragment.DestinationLayoutFragment">
 
     <include
         layout="@layout/top_list"
