+++ app/src/main/res/layout/home_layout.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".HomeLayoutFragment">
+    tools:context=".fragment.HomeLayoutFragment">
 
     <include
         layout="@layout/top_list"
